Changed VS datatypes to cstd types, unittests passed

+++ Headers/BasicGraph.h
@@ -1,37 +1,38 @@
 #ifndef __BASICGRAPH_H__
 #define __BASICGRAPH_H__
+#include <iostream>
+#include <cstdint>
 #include "Utilities.h"
 #include "TS_Logger.h"
-#include <iostream>
 
 template <class nweightType, class tweightType> class BasicGraph
 {
 private:
-	size_t n_nodes;
-	void update_nnodes(size_t node_idx) { if (n_nodes < node_idx + 1) n_nodes = node_idx; };
+	uint64_t n_nodes;
+	void update_nnodes(uint64_t node_idx) { if (n_nodes < node_idx + 1) n_nodes = node_idx; };
 public:
 	struct t_links_struct{
 		tweightType * weights; // source, then sink; double size node_ids; double n_tlinks
-		size_t * node_ids;
-		size_t n_tlinks;
-		size_t maxcap;
+		uint64_t * node_ids;
+		uint64_t n_tlinks;
+		uint64_t maxcap;
 	} t_links;
 
 	struct n_links_struct{
 		nweightType * weights; // n1->n2, then n1<-n2; same size edge_nodes; double n_nlinks
-		size_t * edge_nodes;
-		size_t n_nlinks;
-		size_t maxcap;
+		uint64_t * edge_nodes;
+		uint64_t n_nlinks;
+		uint64_t maxcap;
 	} n_links;
 	
 
 	BasicGraph();
-	BasicGraph(size_t, size_t , size_t ); //set the n_nodes and n_edges maximum capcity, this cap could be excceded but this will be slow
+	BasicGraph(uint64_t, uint64_t , uint64_t ); //set the n_nodes and n_edges maximum capcity, this cap could be excceded but this will be slow
 	void reset(); 
-	void add_nlink(size_t, size_t, nweightType, nweightType);
-	void add_tlink(size_t, tweightType, tweightType);
-	void get_tlink(size_t, tweightType&, tweightType&);
-	void get_nlink(size_t, size_t &, size_t&, nweightType &, nweightType &);
+	void add_nlink(uint64_t, uint64_t, nweightType, nweightType);
+	void add_tlink(uint64_t, tweightType, tweightType);
+	void get_tlink(uint64_t, tweightType&, tweightType&);
+	void get_nlink(uint64_t, uint64_t &, uint64_t&, nweightType &, nweightType &);
 	
 	t_links_struct get_tlinks();
 	n_links_struct get_nlinks();
@@ -39,7 +40,7 @@ public:
 	void saveToFile_Binary(std::string fname);
 	void saveToMaxflowFile(std::string fname);
 
-	size_t getNumberOfNodes() { return this->n_nodes; }
+	uint64_t getNumberOfNodes() { return this->n_nodes; }
 	double getAllocatedMemSizeInGB();
 	~BasicGraph();
 };
@@ -47,18 +48,18 @@ template <class nweightType, class tweightType>
 void BasicGraph<nweightType, tweightType>::saveToMaxflowFile(std::string output_fname)
 {
 	ofstream output_file(output_fname.c_str(), ios::out);
-	size_t s = 1;
-	size_t t = 2;
+	uint64_t s = 1;
+	uint64_t t = 2;
 
 	output_file << "p max " << this->n_nodes << " " << this->t_links.n_tlinks + this->n_links.n_nlinks << std::endl;
 	output_file << "n 1 s" << std::endl;
 	output_file << "n 2 t" << std::endl;
-	for (size_t link_id = 0; link_id < t_links.n_tlinks; ++link_id)
+	for (uint64_t link_id = 0; link_id < t_links.n_tlinks; ++link_id)
 	{
 		output_file <<"a "<< s << " " << t_links.node_ids[link_id]+3 << " " << t_links.weights[link_id * 2 + 0] << std::endl;
 		output_file <<"a "<< t_links.node_ids[link_id]+3 << " " << t << " " << t_links.weights[link_id * 2 + 1] << std::endl;
 	}
-	for (size_t offset = 0; offset < (n_links.n_nlinks * 2); offset += 2)
+	for (uint64_t offset = 0; offset < (n_links.n_nlinks * 2); offset += 2)
 	{
 		output_file << "a " << n_links.edge_nodes[offset + 0] + 3 << " " << n_links.edge_nodes[offset + 1] + 3 << " " << n_links.weights[offset + 0] << std::endl;
 		output_file << "a " << n_links.edge_nodes[offset + 1] + 3 << " " << n_links.edge_nodes[offset + 0] + 3 << " " << n_links.weights[offset + 1] << std::endl;
@@ -81,23 +82,23 @@ BasicGraph<nweightType, tweightType>::BasicGraph()
 };
 
 template <class nweightType, class tweightType>
-BasicGraph<nweightType, tweightType>::BasicGraph(size_t n_nodes, size_t n_tlinks_maxcap, size_t n_nlinks_maxcap)
+BasicGraph<nweightType, tweightType>::BasicGraph(uint64_t n_nodes, uint64_t n_tlinks_maxcap, uint64_t n_nlinks_maxcap)
 {
 	this->n_nodes = n_nodes;
 	this->t_links.n_tlinks = 0;
 	this->t_links.maxcap = n_tlinks_maxcap;
 	this->t_links.weights = new tweightType[this->t_links.maxcap*2];
-	this->t_links.node_ids = new size_t[this->t_links.maxcap];
+	this->t_links.node_ids = new uint64_t[this->t_links.maxcap];
 
 	this->n_links.n_nlinks = 0;
 	this->n_links.maxcap = n_nlinks_maxcap;
 	this->n_links.weights = new nweightType[this->n_links.maxcap * 2];
-	this->n_links.edge_nodes = new size_t[this->n_links.maxcap * 2];
+	this->n_links.edge_nodes = new uint64_t[this->n_links.maxcap * 2];
 };
 template <class nweightType, class tweightType>
 double BasicGraph<nweightType, tweightType>::getAllocatedMemSizeInGB( void )
 {
-	double size_inbytes = sizeof(tweightType) *this->t_links.maxcap * 2 + sizeof(size_t)*this->t_links.maxcap + sizeof(nweightType) * this->n_links.maxcap * 2 + sizeof(size_t)* this->n_links.maxcap * 2;
+	double size_inbytes = sizeof(tweightType) *this->t_links.maxcap * 2 + sizeof(uint64_t)*this->t_links.maxcap + sizeof(nweightType) * this->n_links.maxcap * 2 + sizeof(uint64_t)* this->n_links.maxcap * 2;
 	return (size_inbytes/1024/1024/1024);
 }
 template <class nweightType, class tweightType>
@@ -112,15 +113,15 @@ template <class nweightType, class tweightType>
 void BasicGraph<nweightType, tweightType>::reset()
 {
 	//memset(this->t_links.weights, 0, sizeof(tweightType)*this->t_links.maxcap*2);
-	//memset(this->t_links.node_ids, 0, sizeof(size_t)*this->t_links.maxcap);
+	//memset(this->t_links.node_ids, 0, sizeof(uint64_t)*this->t_links.maxcap);
 	//memset(this->n_links.weights, 0, sizeof(nweightType)*this->n_links.maxcap * 2);
-	//memset(this->n_links.edge_nodes, 0, sizeof(size_t)*this->n_links.maxcap * 2);
+	//memset(this->n_links.edge_nodes, 0, sizeof(uint64_t)*this->n_links.maxcap * 2);
 	this->t_links.n_tlinks = 0;
 	this->n_links.n_nlinks = 0;
 }
 
 template <class nweightType, class tweightType>
-void BasicGraph<nweightType, tweightType>::add_nlink(size_t node0_idx, size_t node1_idx, nweightType w_0to1, nweightType w_1to0)
+void BasicGraph<nweightType, tweightType>::add_nlink(uint64_t node0_idx, uint64_t node1_idx, nweightType w_0to1, nweightType w_1to0)
 {
 	if (this->n_links.n_nlinks == this->n_links.maxcap) //maxcap will be excceded
 	{
@@ -131,14 +132,14 @@ void BasicGraph<nweightType, tweightType>::add_nlink(size_t node0_idx, size_t no
 		CLNDEL1D(this->n_links.weights);
 		this->n_links.weights = new_weights;
 
-		size_t * new_edge_nodes = new size_t[n_links.maxcap * 4];
-		memcpy(new_edge_nodes, this->n_links.edge_nodes, sizeof(size_t)* this->n_links.n_nlinks * 2);
-		memset(new_edge_nodes + this->n_links.n_nlinks * 2, 0, sizeof(size_t)*(n_links.maxcap * 4 - this->n_links.n_nlinks * 2));
+		uint64_t * new_edge_nodes = new uint64_t[n_links.maxcap * 4];
+		memcpy(new_edge_nodes, this->n_links.edge_nodes, sizeof(uint64_t)* this->n_links.n_nlinks * 2);
+		memset(new_edge_nodes + this->n_links.n_nlinks * 2, 0, sizeof(uint64_t)*(n_links.maxcap * 4 - this->n_links.n_nlinks * 2));
 		CLNDEL1D(this->n_links.edge_nodes);
 		this->n_links.edge_nodes = new_edge_nodes;
 		n_links.maxcap *= 2;
 	}
-	size_t edge_id = this->n_links.n_nlinks * 2;
+	uint64_t edge_id = this->n_links.n_nlinks * 2;
 	this->n_links.edge_nodes[edge_id + 0] = node0_idx;
 	this->n_links.edge_nodes[edge_id + 1] = node1_idx;
 	this->n_links.weights[edge_id + 0] = w_0to1;
@@ -148,7 +149,7 @@ void BasicGraph<nweightType, tweightType>::add_nlink(size_t node0_idx, size_t no
 	update_nnodes(node1_idx);
 }
 template <class nweightType, class tweightType>
-void BasicGraph<nweightType, tweightType>::add_tlink(size_t node_idx, tweightType w_src2node, tweightType w_node2sink)
+void BasicGraph<nweightType, tweightType>::add_tlink(uint64_t node_idx, tweightType w_src2node, tweightType w_node2sink)
 {
 	if (this->t_links.maxcap == this->t_links.n_tlinks) //maxcap will be excceded
 	{
@@ -159,9 +160,9 @@ void BasicGraph<nweightType, tweightType>::add_tlink(size_t node_idx, tweightTyp
 		CLNDEL1D(this->t_links.weights);
 		this->t_links.weights = new_weights;
 
-		size_t * new_node_ids = new size_t[this->t_links.maxcap * 2];
-		memcpy(new_node_ids, this->t_links.node_ids, sizeof(size_t)* this->t_links.maxcap);
-		memset(new_node_ids + this->t_links.maxcap, 0, sizeof(size_t)*(t_links.maxcap * 2 - this->t_links.maxcap));
+		uint64_t * new_node_ids = new uint64_t[this->t_links.maxcap * 2];
+		memcpy(new_node_ids, this->t_links.node_ids, sizeof(uint64_t)* this->t_links.maxcap);
+		memset(new_node_ids + this->t_links.maxcap, 0, sizeof(uint64_t)*(t_links.maxcap * 2 - this->t_links.maxcap));
 		CLNDEL1D(this->t_links.node_ids);
 		this->t_links.node_ids = new_node_ids;
 		this->t_links.maxcap *= 2;
@@ -174,7 +175,7 @@ void BasicGraph<nweightType, tweightType>::add_tlink(size_t node_idx, tweightTyp
 	update_nnodes(node_idx);
 }
 template <class nweightType, class tweightType>
-void  BasicGraph<nweightType, tweightType>::get_tlink(size_t link_id, tweightType& w_src2node, tweightType&w_node2sink)
+void  BasicGraph<nweightType, tweightType>::get_tlink(uint64_t link_id, tweightType& w_src2node, tweightType&w_node2sink)
 {
 	if (link_id >= this->t_links.n_tlinks)
 	{
@@ -187,14 +188,14 @@ void  BasicGraph<nweightType, tweightType>::get_tlink(size_t link_id, tweightTyp
 }
 
 template <class nweightType, class tweightType>
-void BasicGraph<nweightType, tweightType>::get_nlink(size_t edge_idx, size_t & node0_idx, size_t& node1_idx, nweightType& w_0to1, nweightType& w_1to0)
+void BasicGraph<nweightType, tweightType>::get_nlink(uint64_t edge_idx, uint64_t & node0_idx, uint64_t& node1_idx, nweightType& w_0to1, nweightType& w_1to0)
 {
 	if (edge_idx >= this->n_links.n_nlinks)
 	{
 		bgn_log << LogType::ERROR << "BG: edge access violation\n" << end_log;
 		throw exception("Access violation");
 	}
-	size_t lcl_edge_idx = edge_idx * 2;
+	uint64_t lcl_edge_idx = edge_idx * 2;
 	node0_idx = this->n_links.edge_nodes[lcl_edge_idx + 0];
 	node1_idx = this->n_links.edge_nodes[lcl_edge_idx + 0];
 	w_0to1 = this->n_links.weights[lcl_edge_idx + 0];
@@ -216,20 +217,20 @@ typename BasicGraph<nweightType, tweightType>::n_links_struct BasicGraph<nweight
 template <class nweightType, class tweightType>
 void BasicGraph<nweightType, tweightType>::saveToFile(std::string fname)
 {
-	ssize_t dim2[2];
+	int64_t dim2[2];
 	dim2[0] = 2;
 	dim2[1] = this->t_links.n_tlinks;
 	Utils::ArrayInOut<tweightType> tweights(this->t_links.weights, dim2, 2);
 	tweights.save(fname + "_BG_tlink_weights.dat");
 	dim2[0] = 1;
-	Utils::ArrayInOut<size_t> nodes(this->t_links.node_ids, dim2, 2);
+	Utils::ArrayInOut<uint64_t> nodes(this->t_links.node_ids, dim2, 2);
 	tweights.save(fname + "_BG_tlink_nodes.dat");
 	
 	dim2[0] = 2;
 	dim2[1] = this->n_links.n_nlinks;
 	Utils::ArrayInOut<nweightType> nweights(this->n_links.weights, dim2, 2);
 	nweights.save(fname + "_BG_nlink_weights.dat");
-	Utils::ArrayInOut<size_t> edges(this->n_links.edge_nodes, dim2, 2);
+	Utils::ArrayInOut<uint64_t> edges(this->n_links.edge_nodes, dim2, 2);
 	edges.save(fname + "_BG_nlink_nodes.dat");
 
 }
@@ -242,7 +243,7 @@ void BasicGraph<nweightType, tweightType>::saveToFile_Binary(std::string fname)
 	output_file1.close();
 
 	ofstream output_file2((fname + "_tlinks_nodes.dat").c_str(), ios::out | ios::binary);
-	output_file2.write(reinterpret_cast<const char*>(this->t_links.node_ids), sizeof(size_t)*this->t_links.n_tlinks);
+	output_file2.write(reinterpret_cast<const char*>(this->t_links.node_ids), sizeof(uint64_t)*this->t_links.n_tlinks);
 	output_file2.close();
 
 
@@ -251,7 +252,7 @@ void BasicGraph<nweightType, tweightType>::saveToFile_Binary(std::string fname)
 	output_file3.close();
 
 	ofstream output_file4((fname + "_nlinks_nodes.dat").c_str(), ios::out | ios::binary);
-	output_file4.write(reinterpret_cast<const char*>(this->n_links.edge_nodes), sizeof(size_t)* this->n_links.n_nlinks * 2);
+	output_file4.write(reinterpret_cast<const char*>(this->n_links.edge_nodes), sizeof(uint64_t)* this->n_links.n_nlinks * 2);
 	output_file4.close();
 
 }
